Add functionality for AppCenter module.
