+++ DemoAppWinCS/windows/DemoAppWinCS/App.xaml.cs
@@ -1,4 +1,8 @@
-﻿using Microsoft.ReactNative;
+﻿using Microsoft.AppCenter;
+using Microsoft.AppCenter.Analytics;
+using Microsoft.ReactNative;
+using System;
+using Windows.ApplicationModel.Activation;
 
 namespace DemoAppWinCS
 {
@@ -30,6 +34,22 @@ namespace DemoAppWinCS
             //PackageProviders.Add(new DemoAppWinCS.ReactPackageProvider());
 
             InitializeComponent();
+        }
+
+        protected override void OnLaunched(LaunchActivatedEventArgs e)
+        {
+            base.OnLaunched(e);
+            StartAppCenter();
+        }
+
+        private static async void StartAppCenter()
+        {
+            var file = await Windows.Storage.StorageFile.GetFileFromApplicationUriAsync(
+                new Uri("ms-appx:///Assets/app-center-config.json"));
+            var content = await Windows.Storage.FileIO.ReadTextAsync(file);
+            var secretContainer = Windows.Data.Json.JsonObject.Parse(content);
+            var appSecret = secretContainer.GetNamedString("app-secret");
+            AppCenter.Start(appSecret);
         }
     }
 }
